fix(today): abbreviate the header date so nothing has to truncate
The previous commit stopped the avatar being clipped by letting the title
block shrink. It worked, and it moved the damage: the date rendered as
"Friday, Septe…" on a 360dp screen.

The long form was 448px of a 1,080px screen for a date, on a screen already
titled "Today". `short` is ~190px, which puts slack back in the row so the
shrink never actually fires — it stays as the safety net for a locale wider
than the one this was tuned against.

Intl does the abbreviating per locale, so es-PR and pt-BR get theirs rather
than an English guess at one.

+++ apps/mobile/src/app/(app)/index.tsx
@@ -51,8 +51,26 @@ import { formatDate } from '@/lib/date-format';
 
 const MIN_STREAK_FOR_REVIEW = 3;
 
+/**
+ * "Fri, Sep 4" — abbreviated on purpose.
+ *
+ * The long form ("Friday, September 4") measured **448px of a 1,080px screen**
+ * on a OnePlus 8T, which is what pushed the header over its width and clipped
+ * the avatar off the right edge. `flexShrink` on the title block stops the
+ * clipping, but with the long form still in place it simply moved the damage
+ * to the date, which rendered as "Friday, Septe…".
+ *
+ * So the shrink is the SAFETY NET and this is the actual fix: at ~190px the
+ * row has slack again and nothing truncates. `short` also degrades better
+ * across locales than `long` does — es-PR's "viernes, 4 de septiembre" is
+ * wider than the English it was tuned against, and Intl handles the
+ * abbreviation per locale rather than us guessing at one.
+ *
+ * Nothing is lost: the screen is titled "Today", so the year and the full
+ * weekday were never carrying information the user needed here.
+ */
 function todayLabel(locale: Locale): string {
-  return formatDate(new Date(), locale, { weekday: 'long', month: 'long', day: 'numeric' });
+  return formatDate(new Date(), locale, { weekday: 'short', month: 'short', day: 'numeric' });
 }
 
 export default function Today() {
